Started readme, realized this is all just deserialization, rewrote readme and adapted package name - parser itself is not renamed yet
