+++ projects/emfular/src/lib/deserialization/constructor-pointers.ts
