+++ projects/emfular/src/public-api.ts
@@ -4,5 +4,5 @@
 
 export * from './lib/referencing/ref';
 export * from './lib/referencing/referenceable';
-export * from './lib/parsing/constructor-pointers';
-export * from './lib/parsing/parser';
+export * from './lib/deserialization/constructor-pointers';
+export * from './lib/deserialization/parser';
